PREAPPS-823: Queue queries into a seperate non-persisted queue then reply them when the OfflineLink is opened, make more options configurable.
