+++ index.ts
@@ -3,7 +3,7 @@ export { ZimbraBatchClient } from './src/batch-client';
 export { LocalBatchLink } from './src/apollo/local-batch-link';
 export { ZimbraErrorLink } from './src/apollo/zimbra-error-link';
 export { ZimbraInMemoryCache } from './src/apollo/zimbra-in-memory-cache';
-export { QueueMutationLink } from './src/apollo/queue-mutation-link';
+export { OfflineQueueLink } from './src/apollo/offline-queue-link';
 export { SyncOfflineMutation } from './src/apollo/sync-offline-mutation';
 export { batchJsonRequest, jsonRequest } from './src/request';
 import * as types from './types';
